requirements.txt
Django==5.0.1
joblib==1.3.2
numpy==1.26.3
pandas==2.2.0
xgboost==2.0.3
psycopg2-binary~=2.9.9
python-dotenv~=1.0.0